Encrypted passwords and other changes
Most of these edits were made on the Visions For Change Norwich website in 2016, but were never uploaded here. I now forget what exactly all the edits were, and I know there are still some issues with them.

IF YOU ARE INCLINED TO USE THIS CODE, PLEASE TEST AND CHECK CAREFULLY FOR BUGS.